Group emails will send at once time, refactored the API Sender
